Rename: mapping/entity 삭제 및 하위 클래스들 이동
