+++ src/main/java/com/otakumap/domain/event/entity/Event.java
@@ -6,7 +6,7 @@ import com.otakumap.domain.event.entity.enums.Genre;
 import com.otakumap.domain.eventLocation.entity.EventLocation;
 import com.otakumap.domain.event_like.entity.EventLike;
 import com.otakumap.domain.image.entity.Image;
-import com.otakumap.domain.mapping.entity.EventAnimation;
+import com.otakumap.domain.mapping.EventAnimation;
 import com.otakumap.global.common.BaseEntity;
 import jakarta.persistence.*;
 import lombok.*;
