+++ src/main/java/com/otakumap/domain/reviews/repository/ReviewRepositoryCustom.java
@@ -3,6 +3,6 @@ package com.otakumap.domain.reviews.repository;
 import com.otakumap.domain.reviews.dto.ReviewResponseDTO;
 import org.springframework.data.domain.Page;
 
-public interface ReviewsRepositoryCustom {
+public interface ReviewRepositoryCustom {
     Page<ReviewResponseDTO.SearchedReviewPreViewDTO> getReviewsByKeyword(String keyword, int page, int size, String sort);
 }
