+++ src/main/java/com/otakumap/domain/reviews/service/ReviewQueryServiceImpl.java
@@ -1,7 +1,7 @@
 package com.otakumap.domain.reviews.service;
 
 import com.otakumap.domain.reviews.dto.ReviewResponseDTO;
-import com.otakumap.domain.reviews.repository.ReviewsRepositoryCustom;
+import com.otakumap.domain.reviews.repository.ReviewRepositoryCustom;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.stereotype.Service;
@@ -12,11 +12,11 @@ import org.springframework.transaction.annotation.Transactional;
 @Transactional(readOnly = true)
 public class ReviewQueryServiceImpl implements ReviewQueryService {
 
-    private final ReviewsRepositoryCustom reviewsRepositoryCustom;
+    private final ReviewRepositoryCustom reviewRepositoryCustom;
 
     @Override
     public Page<ReviewResponseDTO.SearchedReviewPreViewDTO> searchReviewsByKeyword(String keyword, int page, int size, String sort) {
 
-        return reviewsRepositoryCustom.getReviewsByKeyword(keyword, page, size, sort);
+        return reviewRepositoryCustom.getReviewsByKeyword(keyword, page, size, sort);
     }
 }
